Android: Handle nested notification bundle
Setup: A backend is sending push notifications via GCM using XMPP. The
message uses the notification payload format, which is a JSON object
(see
https://developers.google.com/cloud-messaging/xmpp-server-ref#downstream-xmpp-messages-json).

Issue: On an Android device, the notification payload arrives as a
nested Bundle with key "notification", within the outer "notification"
Bundle that is passed to notifyNotification(). In its current form,
convertJSON() doesn't convert this nested Bundle into a string, and as
such, the notification payload is missing from the notification that
is passed to the onNotification() callback.

Fix: In convertJSON(), call itself recursively to convert nested
Bundles into JSON. Change the function's return type to JSONObject.