Update README.md
Added info to help with a gotcha we encountered whilst using this package.  Lacking the super call meant that the extras from the notification were ignored.

We couldn't get onNotifications to fire on Android local scheduled notifications and we found that during an installation of another module, namely React Native Branch Deep Linking, onNewIntent was overridden without adding the super.onNewIntent() bit.  This and possibly other RN modules could be leading to some of the difficulties with onNotification not firing.